Separated Scanner from the Server and Added a Client class
Removed the Scanner hardbuild into the Server class and moved it into a Client Class. Add and Search server methods working but i'm not sure if it can run with more than one client yet.

Server class might become Gateway class or might need to create the 2nd to handle multiple clients and that's the class that calls the methods in Server

To see how to run everything, read READ_ME.md